% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo2L.R
\name{phylo2L}
\alias{phylo2L}
\title{Convert bd phylo to L table}
\usage{
phylo2L(emdata)
}
\arguments{
\item{emdata}{a phylogeny, of type \code{\link[ape]{phylo}}}
}
\value{
the L table
}
\description{
Convert bd phylo to L table. Don't use for mbd.
}
\author{
Xu Liang
}
